backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
sqlalchemy>=2.0.9
psycopg2-binary>=2.9.6
celery>=5.2.7
redis>=4.5.4
openai>=0.27.4
python-dotenv>=1.0.0
pgvector>=0.1.8
httpx>=0.24.0
pytest>=7.3.1
pytest-asyncio>=0.21.0
python-multipart>=0.0.6
